fix m1 controller error
